Some moar changes:
 - Add pascal extensions
 - Switch to widths in doubles
 - Implement saving resized/reordered columns per protocol
 - Fix a small bug related to that
 - Don't call indexer multiple times in UpdateViews
 - Fix X-thread violation while updating items
 - Fix parsing queue
 - Fix ViewerItemCollection.Update(ViewerItem)
 - Make ViewerItemColleciton optionally implement INotifyPropertyChanged

+++ Kamilla/Network/Protocols/Protocol.cs
@@ -46,7 +46,7 @@ namespace Kamilla.Network.Protocols
         /// Gets widths of columns when packets of the current
         /// <see cref="Kamilla.Network.Protocols.Protocol"/> are displayed in a ListView.
         /// </summary>
-        public abstract int[] ListViewColumnWidths { get; }
+        public abstract double[] ListViewColumnWidths { get; }
 
         protected virtual PacketParser InternalCreateParser(ViewerItem item)
         {
